requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.1.4
django-countries==7.0
django-filter==0.9.2
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
pytz==2024.1
requests==2.31.0
six==1.15.0
sqlparse==0.5.0
stripe==2.55.2
urllib3==1.26.3
wrapt==1.12.1
